If IConsumerErrorStrategy throws Exception, message remains unacked (#803)
* Fix #801

* Add tests for HandlerRunner when user handler is failed or cancelled.

* Fix tests
